api/requirements.txt
# Caption-It API - Production Dependencies
# Python 3.10+

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Config
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client (for HF API & URL fetching)
httpx>=0.26.0

# Hugging Face (for local model inference)
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0

# Image processing
Pillow>=10.0.0

# Optional: safetensors for model loading
safetensors>=0.4.0
